move registerTest() from RegistrationServiceTest to RegistrationServiceTestNew
