+++ service-registration/src/test/java/sg/ncl/service/registration/Util.java
@@ -63,6 +63,7 @@ public class Util {
     public static UserEntity getUserEntity() {
         final UserEntity userEntity = new UserEntity();
         userEntity.setApplicationDate(ZonedDateTime.now());
+        userEntity.setVerificationKey(RandomStringUtils.randomAlphanumeric(20));
 
         final UserDetailsEntity userDetailsEntity = new UserDetailsEntity();
         userDetailsEntity.setFirstName(RandomStringUtils.randomAlphabetic(20));
